Addressing PR feedback and adding span debugger display string.
