+++ src/mscorlib/shared/System/ReadOnlySpan.cs
@@ -15,6 +15,7 @@ namespace System
     /// ReadOnlySpan represents a contiguous region of arbitrary memory. Unlike arrays, it can point to either managed
     /// or native memory, or to memory allocated on the stack. It is type- and memory-safe.
     /// </summary>
+    [DebuggerDisplay("{DebuggerDisplay,nq}")]
     [IsReadOnly]
     [IsByRefLike]
     [NonVersionable]
@@ -117,6 +118,9 @@ namespace System
             _pointer = new ByReference<T>(ref ptr);
             _length = length;
         }
+
+        //Debugger Display = {T[length]}
+        private string DebuggerDisplay => string.Format("{{{0}[{1}]}}", typeof(T).Name, _length);
 
         /// <summary>
         /// Returns a reference to the 0th element of the Span. If the Span is empty, returns a reference to the location where the 0th element
